==> main/03_construct_figuretable.Rmd <==
---
title: "03_construct_figuretable"
author: "Suguru Otani"
date: "2024-01-26"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# load {.tabset}

```{r, echo=FALSE,results = "asis"}
rm(list = ls())
library(magrittr)
library(ggplot2)
library(modelsummary) # for N in datasummary

utmd_output_hello_work_data_yearly <-
  readRDS(file = here::here("output/utmd_output_hello_work_data_yearly.rds")) %>% 
  dplyr::rename(
    `unemployed`= candidate_count,
    `vacancies` = position_count,
    `hire` = hire_count,
    `dlog(M)/ dlog(V)` = hire_elasticity_vacancy,
    `dlog(M)/ dlog(AU)` = hire_elasticity_efficiency_unemployed
  )

utmd_output_hello_work_data_full_time_monthly <-
  readRDS(file = here::here("output/utmd_output_hello_work_data_full_time_monthly.rds")) %>% 
  dplyr::rename(
    `unemployed`= candidate_count,
    `vacancies` = position_count,
    `hire` = hire_count,
    `dlog(M)/ dlog(V)` = hire_elasticity_vacancy,
    `dlog(M)/ dlog(AU)` = hire_elasticity_efficiency_unemployed
  )
utmd_output_hello_work_data_part_time_monthly <-
  readRDS(file = here::here("output/utmd_output_hello_work_data_part_time_monthly.rds")) %>% 
  dplyr::rename(
    `unemployed`= candidate_count,
    `vacancies` = position_count,
    `hire` = hire_count,
    `dlog(M)/ dlog(V)` = hire_elasticity_vacancy,
    `dlog(M)/ dlog(AU)` = hire_elasticity_efficiency_unemployed
  )
utmd_output_hello_work_data_part_and_full_time_monthly <-
  readRDS(file = here::here("output/utmd_output_hello_work_data_part_and_full_time_monthly.rds")) %>% 
  dplyr::rename(
    `unemployed`= candidate_count,
    `vacancies` = position_count,
    `hire` = hire_count,
    `dlog(M)/ dlog(V)` = hire_elasticity_vacancy,
    `dlog(M)/ dlog(AU)` = hire_elasticity_efficiency_unemployed
  )
utmd_output_hello_work_data_pooling_part_and_full_time_monthly <-
  readRDS(file = here::here("output/utmd_output_hello_work_data_pooling_part_and_full_time_monthly.rds")) %>% 
  dplyr::rename(
    `unemployed`= candidate_count,
    `vacancies` = position_count,
    `hire` = hire_count,
    `dlog(M)/ dlog(V)` = hire_elasticity_vacancy,
    `dlog(M)/ dlog(AU)` = hire_elasticity_efficiency_unemployed
  )

```

## Figure 1: Residual plot {.tabset}



### Month-level part-time and full-time

```{r,echo=FALSE,results = "asis"}
x <-
  ggplot(
    utmd_output_hello_work_data_part_and_full_time_monthly,
     aes(
       x = residual_v_on_u,
       y = residual_a_on_u
       )
     ) +
  #geom_line(size = 0.3) +
  geom_point() +
  #geom_text(aes(label = time), hjust = 0.5, vjust = -0.5, size = 3) +
  theme_classic() +
  labs(
    #title = "Beveridge Curve",
    x = "Residual of V on U",
    y = "Residual of A on U"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.caption = element_text(hjust = 0, size = 8)
  ) +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/residual_plot_month_aggregate.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```




### Month-level pooling full-time and part-time 

```{r,echo=FALSE,results = "asis"}
x <-
  ggplot(
    utmd_output_hello_work_data_pooling_part_and_full_time_monthly,
     aes(
       x = residual_v_on_u,
       y = residual_a_on_u,
       color = job_type
       )
     ) +
  #geom_line(size = 0.3) +
  geom_point() +
  #geom_text(aes(label = time), hjust = 0.5, vjust = -0.5, size = 3) +
  theme_classic() +
  labs(
    #title = "Beveridge Curve",
    x = "Residual of V on U",
    y = "Residual of A on U"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.caption = element_text(hjust = 0, size = 8)
  ) +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  labs(colour = "") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/residual_plot_month_pooling_part_and_full_time.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```

### Year-level

```{r,echo=FALSE,results = "asis"}
x <-
  ggplot(
    utmd_output_hello_work_data_yearly,
     aes(
       x = residual_v_on_u,
       y = residual_a_on_u
       )
     ) +
  #geom_line(size = 0.3) +
  geom_point() +
  #geom_text(aes(label = time), hjust = 0.5, vjust = -0.5, size = 3) +
  theme_classic() +
  labs(
    #title = "Beveridge Curve",
    x = "Residual of V on U",
    y = "Residual of A on U"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.caption = element_text(hjust = 0, size = 8)
  ) +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/residual_plot_year.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```


## Figure 2: Month-level aggregate results 1973-2024 {.tabset}

the match elasticity with respect to unemployment is in the range 0.5–0.7.

### U, V, and V/U

```{r,echo=FALSE,results = "asis"}
temp <-
  utmd_output_hello_work_data_part_and_full_time_monthly %>% 
  dplyr::select(
    `unemployed`,
    `vacancies`,
    time
  ) %>% 
  tidyr::gather(
    type,
    count,
    - time
  )
temp2 <-
  utmd_output_hello_work_data_part_and_full_time_monthly %>% 
  dplyr::mutate(
    `V/U` =
      `vacancies`/`unemployed`
  ) %>% 
  dplyr::select(
    `V/U`,
    time
  ) %>% 
  tidyr::gather(
    type,
    count,
    - time
  )
x <- 
  ggplot(
    temp,
    aes(
      x = time,
      y = count/1000000,
      color = type, 
      group = type
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  theme_classic() + 
  geom_line(
    data = temp2,
    aes(
      x = time, 
      y = count
      ), 
    alpha = 0.8, 
    size = 0.3,
    linetype = "dashed"
      ) + 
  scale_y_continuous(
    name = "Count (million) (solid)",
    sec.axis = sec_axis(
      ~.*1, 
      name = "V/U (dashed)"
      )
    ) +
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  theme(axis.text.x = element_text(angle=60, vjust=1, hjust=1, size=3.0)) +
  xlab("") +
  labs(colour = "") +
  #ylab("Count (million)") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/unemployed_vacancy_month_aggregate.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)


```


### H

```{r,echo=FALSE,results = "asis"}
x <- 
  ggplot(
    utmd_output_hello_work_data_part_and_full_time_monthly,
    aes(
      x = time,
      y = `hire`/1000000
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  # geom_point(
  #   aes(
  #     shape = industry_group_name,
  #     color = industry_group_name
  #     )
  #   ) +
  theme_classic() + 
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  xlab("") +
  ylab("Count (million)") +
  labs(colour = "") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/hire_month_aggregate.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)

```

### U and V relationship

```{r,echo=FALSE,results = "asis"}
x <-
  ggplot(
    utmd_output_hello_work_data_part_and_full_time_monthly %>% 
      dplyr::mutate(
        time = ifelse(
          month == "01-01",
          year,
          NA
        )
      ),
     aes(
       x = unemployed/1000000,
       y = vacancies/1000000
       )
     ) +
  geom_line(size = 0.3) +
  geom_point(size = 0.3, alpha = 0.3) +
  geom_text(aes(label = time), hjust = 0.5, vjust = -0.5, size = 2) +
  theme_classic() +
  labs(
    #title = "Beveridge Curve",
    x = "U (mil)",
    y = "V (mil)"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.caption = element_text(hjust = 0, size = 8)
  ) +
  labs(colour = "") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/unemployed_vacancies_berveridge_month_aggregate.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```

### Job finding rate and worker finding rate

```{r,echo=FALSE,results = "asis"}
temp <-
  utmd_output_hello_work_data_part_and_full_time_monthly %>% 
  dplyr::mutate(
    `M/U` =
      `hire`/`unemployed`,
    `M/V` =
      `hire`/`vacancies`
  ) %>% 
  dplyr::select(
    `M/U`,
    `M/V`,
    time
  ) %>% 
  tidyr::gather(
    type,
    count,
    - time
  )

x <- 
  ggplot(
    temp,
    aes(
      x = time,
      y = count,
      color = type, 
      group = type
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  theme_classic() + 
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  #theme(axis.text.x = element_text(angle=60, vjust=1, hjust=1, size=3.0)) +
  xlab("") +
  ylab("") +
  labs(colour = "") +
  #ylab("Count (million)") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/job_finding_rate_worker_finding_rate_month_aggregate.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```


### Matching Efficiency A

```{r,echo=FALSE,results = "asis"}
x <- 
  ggplot(
    utmd_output_hello_work_data_part_and_full_time_monthly,
    aes(
      x = time,
      y = efficiency_implied
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  # geom_point(
  #   aes(
  #     shape = industry_group_name,
  #     color = industry_group_name
  #     )
  #   ) +
  geom_hline(yintercept = 100, linetype = "longdash") +
  theme_classic() + 
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8))+
  #theme(axis.text.x = element_text(angle=60, vjust=1, hjust=1, size=3.0)) +
  xlab("") +
  ylab("A (normalized to 1972 Jan)") +
  labs(colour = "") +
  #geom_vline(xintercept = 2002, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/matching_efficiency_month_aggregate.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```

### Elasticity dlogM/dlogU,dlogM/dlogV

```{r,echo=FALSE,results = "asis"}
temp <-
  utmd_output_hello_work_data_part_and_full_time_monthly %>% 
  dplyr::select(
    `dlog(M)/ dlog(V)`,
    `dlog(M)/ dlog(AU)`,
    time
  ) %>% 
  tidyr::gather(
    type,
    count,
    - time
  )
x <- 
  ggplot(
    temp,
    aes(
      x = time,
      y = count,
      color = type, 
      group = type
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  # geom_point(
  #   aes(
  #     shape = industry_group_name,
  #     color = industry_group_name
  #     )
  #   ) +
  #geom_hline(yintercept = 100, linetype = "longdash") +
  theme_classic() + 
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8))+
  xlab("") +
  ylab("log(M) w.r.t. log(AU) and log(V)") +
  labs(colour = "") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/elasticity_month_aggregate.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
x
```





### The labor market tightness ($V/U$) and Efficiency ($A$)

```{r,echo=FALSE,results = "asis"}
x <-
  ggplot(
    utmd_output_hello_work_data_part_and_full_time_monthly,
     aes(
       x = log(vacancies/unemployed),
       y = efficiency_implied
       )
     ) +
  #geom_line(size = 0.3) +
  geom_point(
    alpha = 0.8, 
    size = 0.5
  ) +
  #geom_text(aes(label = time), hjust = 0.5, vjust = -0.5, size = 3) +
  theme_classic() +
  labs(
    #title = "Beveridge Curve",
    x = "log(V/U)",
    y = "A"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.caption = element_text(hjust = 0, size = 8)
  ) +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/efficiency_tightness_plot_month_aggregate.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```

### The job finding rate ($H/U$) , filling rate ($H/V$) and Efficiency ($A$)

```{r,echo=FALSE,results = "asis"}
temp <-
  utmd_output_hello_work_data_part_and_full_time_monthly %>% 
  dplyr::mutate(
    `log(M/U)` =
      log(`hire`/`unemployed`),
    `log(M/V)` =
      log(`hire`/`vacancies`)
  ) %>% 
  dplyr::select(
    `log(M/U)`,
    `log(M/V)`,
    efficiency_implied
  ) %>% 
  tidyr::gather(
    type,
    count,
    - efficiency_implied
  )

x <-
  ggplot(
    temp,
    aes(
      x = count,
      y = efficiency_implied,
      color = type, 
      group = type
      )
    ) +
  #geom_line(size = 0.5) +
  geom_point(
    alpha = 0.8, 
    size = 0.3
  ) +
  #geom_text(aes(label = time), hjust = 0.5, vjust = -0.5, size = 3) +
  theme_classic() +
  labs(
    #title = "Beveridge Curve",
    x = "log(M/U), log(M/V)",
    y = "A"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.caption = element_text(hjust = 0, size = 8)
  ) +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  labs(colour = "") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/job_finding_rate_efficiency_plot_month_aggregate.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```



## Figure 3: Month-level full-time and part-time trends in 1972-2024 {.tabset}

the match elasticity with respect to unemployment is in the range 0.5–0.7.

### U, V, and V/U

```{r,echo=FALSE,results = "asis"}
temp <-
  utmd_output_hello_work_data_pooling_part_and_full_time_monthly %>% 
  dplyr::filter(job_type == "full-time") %>% 
  dplyr::select(
    `unemployed`,
    `vacancies`,
    time
  ) %>% 
  dplyr::rename(
   `unemployed (full)` = `unemployed`,
   `vacancies (full)` =`vacancies`
  ) %>% 
  tidyr::gather(
    type,
    count,
    - time
  )
temp2 <-
  utmd_output_hello_work_data_pooling_part_and_full_time_monthly %>% 
  dplyr::filter(job_type == "full-time") %>% 
  dplyr::mutate(
    `V/U` =
      `vacancies`/`unemployed`
  ) %>% 
  dplyr::select(
    `V/U`,
    time
  ) %>% 
  dplyr::rename(
   `V/U (full)` = `V/U`
  ) %>% 
  tidyr::gather(
    type,
    count,
    - time
  )
temp3 <-
  utmd_output_hello_work_data_pooling_part_and_full_time_monthly %>% 
  dplyr::filter(job_type == "part-time") %>% 
  dplyr::select(
    `unemployed`,
    `vacancies`,
    time
  ) %>% 
  dplyr::rename(
   `unemployed (part)` = `unemployed`,
   `vacancies (part)` =`vacancies`
  ) %>% 
  tidyr::gather(
    type,
    count,
    - time
  )
temp <-
  rbind(
    temp,
    temp3
  )


temp4 <-
  utmd_output_hello_work_data_pooling_part_and_full_time_monthly %>% 
  dplyr::filter(job_type == "part-time") %>% 
  dplyr::mutate(
    `V/U` =
      `vacancies`/`unemployed`
  ) %>% 
  dplyr::select(
    `V/U`,
    time
  ) %>% 
  dplyr::rename(
   `V/U (part)` = `V/U`
  ) %>% 
  tidyr::gather(
    type,
    count,
    - time
  )
temp2 <-
  rbind(
    temp2,
    temp4
  )
x <- 
  ggplot(
    temp,
    aes(
      x = time,
      y = count/1000000,
      color = type, 
      group = type
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  theme_classic() + 
  geom_line(
    data = temp2,
    aes(
      x = time, 
      y = count
      ), 
    alpha = 0.8, 
    size = 0.3,
    linetype = "dashed"
      ) + 
  scale_y_continuous(
    name = "Count (million) (solid)",
    sec.axis = sec_axis(
      ~.*1, 
      name = "V/U (dashed)"
      )
    ) +
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  theme(axis.text.x = element_text(angle=60, vjust=1, hjust=1, size=3.0)) +
  xlab("") +
  labs(colour = "") +
  #ylab("Count (million)") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/unemployed_vacancy_month_full_time_part_time.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)


```


### H

```{r,echo=FALSE,results = "asis"}
temp <-
  utmd_output_hello_work_data_pooling_part_and_full_time_monthly %>% 
  dplyr::filter(job_type == "full-time") %>% 
  dplyr::select(
    `hire`,
    time
  ) %>% 
  dplyr::rename(
   `hire (full)` = `hire`
  ) %>% 
  tidyr::gather(
    type,
    count,
    - time
  )
temp2 <-
  utmd_output_hello_work_data_pooling_part_and_full_time_monthly %>% 
  dplyr::filter(job_type == "part-time") %>% 
  dplyr::select(
    `hire`,
    time
  )  %>% 
  dplyr::rename(
   `hire (part)` = `hire`
  ) %>% 
  tidyr::gather(
    type,
    count,
    - time
  )
temp <-
  rbind(
    temp,
    temp2
  )
x <- 
  ggplot(
    temp,
    aes(
      x = time,
      y = count/1000000,
      color = type, 
      group = type
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  # geom_point(
  #   aes(
  #     shape = industry_group_name,
  #     color = industry_group_name
  #     )
  #   ) +
  theme_classic() + 
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  xlab("") +
  ylab("Count (million)") +
  labs(colour = "") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/hire_month_full_time_part_time.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)

```

### U and V relationship

```{r,echo=FALSE,results = "asis"}
temp <-
  utmd_output_hello_work_data_pooling_part_and_full_time_monthly %>% 
  dplyr::filter(job_type == "full-time") %>% 
  dplyr::select(
    month,
    year,
    unemployed,
    vacancies
  ) %>% 
  dplyr::mutate(
    type = "full"
  )
temp2 <-
  utmd_output_hello_work_data_pooling_part_and_full_time_monthly %>% 
  dplyr::filter(job_type == "part-time") %>% 
    dplyr::select(
    month,
    year,
    unemployed,
    vacancies
  ) %>% 
  dplyr::mutate(
    type = "part"
  )
temp <-
  rbind(
    temp,
    temp2
    )


x <-
  ggplot(
    temp %>% 
      dplyr::mutate(
        time = ifelse(
          month == "01-01",
          year,
          NA
        )
      ),
     aes(
       x = unemployed/1000000,
       y = vacancies/1000000,
      color = type, 
      group = type
       )
     ) +
  geom_line(size = 0.3, alpha = 0.8) +
  geom_point(size = 0.3, alpha = 0.3) +
  geom_text(aes(label = time), hjust = 0.5, vjust = -0.5, size = 2) +
  theme_classic() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  labs(
    #title = "Beveridge Curve",
    x = "U (mil)",
    y = "V (mil)"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.caption = element_text(hjust = 0, size = 8)
  ) +
  labs(colour = "") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/unemployed_vacancies_berveridge_month_full_time_part_time.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```

### Job finding rate and worker finding rate

```{r,echo=FALSE,results = "asis"}
temp <-
  utmd_output_hello_work_data_pooling_part_and_full_time_monthly %>% 
  dplyr::filter(job_type == "full-time") %>% 
  dplyr::mutate(
    `M/U (full)` =
      `hire`/`unemployed`,
    `M/V (full)` =
      `hire`/`vacancies`
  ) %>% 
  dplyr::select(
    `M/U (full)`,
    `M/V (full)`,
    time
  ) %>% 
  tidyr::gather(
    type,
    count,
    - time
  )

temp2 <-
  utmd_output_hello_work_data_pooling_part_and_full_time_monthly %>% 
  dplyr::filter(job_type == "part-time") %>% 
  dplyr::mutate(
    `M/U (part)` =
      `hire`/`unemployed`,
    `M/V (part)` =
      `hire`/`vacancies`
  ) %>% 
  dplyr::select(
    `M/U (part)`,
    `M/V (part)`,
    time
  ) %>% 
  tidyr::gather(
    type,
    count,
    - time
  )
temp <-
  rbind(
    temp,
    temp2
  )

x <- 
  ggplot(
    temp,
    aes(
      x = time,
      y = count,
      color = type, 
      group = type
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  theme_classic() + 
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  #theme(axis.text.x = element_text(angle=60, vjust=1, hjust=1, size=3.0)) +
  xlab("") +
  ylab("") +
  labs(colour = "") +
  #ylab("Count (million)") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/job_finding_rate_worker_finding_rate_month_full_time_part_time.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```

## Figure 4: (Continued) Month-level full-time and part-time trends in 1972-2024 {.tabset}

### Matching Efficiency A

```{r,echo=FALSE,results = "asis"}
# temp <-
#   utmd_output_hello_work_data_full_time_monthly %>% 
#   dplyr::select(
#     time,
#     efficiency_implied
#   ) %>% 
#   dplyr::mutate(
#     type = "full"
#   )
# temp2 <-
#   utmd_output_hello_work_data_part_time_monthly %>% 
#   dplyr::select(
#     time,
#     efficiency_implied
#   ) %>% 
#   dplyr::mutate(
#     type = "part"
#   )
# temp <-
#   rbind(
#     temp,
#     temp2
#   )

x <- 
  ggplot(
    utmd_output_hello_work_data_pooling_part_and_full_time_monthly,
    aes(
      x = time,
      y = efficiency_implied,
      color = job_type, 
      group = job_type
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  # geom_point(
  #   aes(
  #     shape = industry_group_name,
  #     color = industry_group_name
  #     )
  #   ) +
  geom_hline(yintercept = 100, linetype = "longdash") +
  theme_classic() + 
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8))+
  #theme(axis.text.x = element_text(angle=60, vjust=1, hjust=1, size=3.0)) +
  xlab("") +
  ylab("A (normalized to 1972 Jan full-time)") +
  labs(colour = "") +
  #geom_vline(xintercept = 2002, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/matching_efficiency_month_full_time_part_time.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```



### Elasticity dlogM/dlogU,dlogM/dlogV

```{r,echo=FALSE,results = "asis"}
temp <-
  utmd_output_hello_work_data_pooling_part_and_full_time_monthly %>% 
  dplyr::filter(job_type == "full-time") %>% 
  dplyr::select(
    `dlog(M)/ dlog(V)`,
    `dlog(M)/ dlog(AU)`,
    time
  ) %>% 
  dplyr::rename(
    `dlog(M)/ dlog(V) (full)` = `dlog(M)/ dlog(V)`,
    `dlog(M)/ dlog(AU) (full)` = `dlog(M)/ dlog(AU)`
  ) %>% 
  tidyr::gather(
    job_type,
    count,
    - time
  )
temp2 <-
  utmd_output_hello_work_data_pooling_part_and_full_time_monthly %>% 
  dplyr::filter(job_type == "part-time") %>% 
  dplyr::select(
    `dlog(M)/ dlog(V)`,
    `dlog(M)/ dlog(AU)`,
    time
  ) %>% 
  dplyr::rename(
    `dlog(M)/ dlog(V) (part)` = `dlog(M)/ dlog(V)`,
    `dlog(M)/ dlog(AU) (part)` = `dlog(M)/ dlog(AU)`
  ) %>% 
  tidyr::gather(
    job_type,
    count,
    - time
  )
temp <-
  rbind(
    temp,
    temp2
  )

x <- 
  ggplot(
    temp,
    aes(
      x = time,
      y = count,
      color = job_type, 
      group = job_type
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  # geom_point(
  #   aes(
  #     shape = industry_group_name,
  #     color = industry_group_name
  #     )
  #   ) +
  #geom_hline(yintercept = 100, linetype = "longdash") +
  theme_classic() + 
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8))+
  xlab("") +
  ylab("log(M) w.r.t. log(AU) and log(V)") +
  labs(colour = "") +
  guides(color = guide_legend(nrow = 2)) +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/elasticity_month_full_time_part_time.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```


### The labor market tightness ($V/U$) and Efficiency ($A$) (full-time)

```{r,echo=FALSE,results = "asis"}
x <-
  ggplot(
    utmd_output_hello_work_data_pooling_part_and_full_time_monthly %>% 
      dplyr::filter(job_type == "full-time"),
     aes(
       x = log(vacancies/unemployed),
       y = efficiency_implied
       )
     ) +
  #geom_line(size = 0.3) +
  geom_point(
    alpha = 0.8, 
    size = 0.5
  ) +
  #geom_text(aes(label = time), hjust = 0.5, vjust = -0.5, size = 3) +
  theme_classic() +
  labs(
    #title = "Beveridge Curve",
    x = "log(V/U)",
    y = "A"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.caption = element_text(hjust = 0, size = 8)
  ) +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/efficiency_tightness_plot_month_full_time.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```




### The labor market tightness ($V/U$) and Efficiency ($A$) (part-time)

```{r,echo=FALSE,results = "asis"}
x <-
  ggplot(
    utmd_output_hello_work_data_pooling_part_and_full_time_monthly %>% 
      dplyr::filter(job_type == "part-time"),
     aes(
       x = log(vacancies/unemployed),
       y = efficiency_implied
       )
     ) +
  #geom_line(size = 0.3) +
  geom_point(
    alpha = 0.8, 
    size = 0.5
  ) +
  #geom_text(aes(label = time), hjust = 0.5, vjust = -0.5, size = 3) +
  theme_classic() +
  labs(
    #title = "Beveridge Curve",
    x = "log(V/U)",
    y = "A"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.caption = element_text(hjust = 0, size = 8)
  ) +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/efficiency_tightness_plot_month_part_time.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```

### The job finding rate ($H/U$) , filling rate ($H/V$) and Efficiency ($A$) (full-time)

```{r,echo=FALSE,results = "asis"}
temp <-
  utmd_output_hello_work_data_pooling_part_and_full_time_monthly %>% 
  dplyr::filter(job_type == "full-time") %>% 
  dplyr::mutate(
    `log(M/U)` =
      log(`hire`/`unemployed`),
    `log(M/V)` =
      log(`hire`/`vacancies`)
  ) %>% 
  dplyr::select(
    `log(M/U)`,
    `log(M/V)`,
    efficiency_implied
  ) %>% 
  tidyr::gather(
    type,
    count,
    - efficiency_implied
  )

x <-
  ggplot(
    temp,
    aes(
      x = count,
      y = efficiency_implied,
      color = type, 
      group = type
      )
    ) +
  #geom_line(size = 0.3) +
  geom_point(
    alpha = 0.8, 
    size = 0.5
  ) +
  #geom_text(aes(label = time), hjust = 0.5, vjust = -0.5, size = 3) +
  theme_classic() +
  labs(
    #title = "Beveridge Curve",
    x = "log(M/U), log(M/V)",
    y = "A"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.caption = element_text(hjust = 0, size = 8)
  ) +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  labs(colour = "") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/job_finding_rate_efficiency_plot_month_full_time.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```




the match elasticity with respect to unemployment is in the range 0.5–0.7.









### The job finding rate ($H/U$) , filling rate ($H/V$) and Efficiency ($A$)  (part-time)

```{r,echo=FALSE,results = "asis"}
temp <-
  utmd_output_hello_work_data_pooling_part_and_full_time_monthly %>% 
  dplyr::filter(job_type == "part-time") %>% 
  dplyr::mutate(
    `log(M/U)` =
      log(`hire`/`unemployed`),
    `log(M/V)` =
      log(`hire`/`vacancies`)
  ) %>% 
  dplyr::select(
    `log(M/U)`,
    `log(M/V)`,
    efficiency_implied
  ) %>% 
  tidyr::gather(
    type,
    count,
    - efficiency_implied
  )

x <-
  ggplot(
    temp,
    aes(
      x = count,
      y = efficiency_implied,
      color = type, 
      group = type
      )
    ) +
  #geom_line(size = 0.3) +
  geom_point(
    alpha = 0.8, 
    size = 0.5
  ) +
  #geom_text(aes(label = time), hjust = 0.5, vjust = -0.5, size = 3) +
  theme_classic() +
  labs(
    #title = "Beveridge Curve",
    x = "log(M/U), log(M/V)",
    y = "A"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.caption = element_text(hjust = 0, size = 8)
  ) +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  labs(colour = "") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/job_finding_rate_efficiency_plot_month_part_time.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```



## Figure 5: Year-level trends in 1966-2023 {.tabset}

the match elasticity with respect to unemployment is in the range 0.5–0.7.

### U, V, and V/U

```{r,echo=FALSE,results = "asis"}
temp <-
  utmd_output_hello_work_data_yearly %>% 
  dplyr::select(
    `unemployed`,
    `vacancies`,
    time
  ) %>% 
  tidyr::gather(
    type,
    count,
    - time
  )
temp2 <-
  utmd_output_hello_work_data_yearly %>% 
  dplyr::mutate(
    `V/U` =
      `vacancies`/`unemployed`
  ) %>% 
  dplyr::select(
    `V/U`,
    time
  ) %>% 
  tidyr::gather(
    type,
    count,
    - time
  )
x <- 
  ggplot(
    temp,
    aes(
      x = time,
      y = count/1000000,
      color = type, 
      group = type
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  theme_classic() + 
  geom_line(
    data = temp2,
    aes(
      x = time, 
      y = count
      ), 
    alpha = 0.8, 
    size = 0.3,
    linetype = "dashed"
      ) + 
  scale_y_continuous(
    name = "Count (million) (solid)",
    sec.axis = sec_axis(
      ~.*1, 
      name = "V/U (dashed)"
      )
    ) +
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  theme(axis.text.x = element_text(angle=60, vjust=1, hjust=1, size=3.0)) +
  xlab("") +
  labs(colour = "") +
  #ylab("Count (million)") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/unemployed_vacancy_year.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)

```


### H

```{r,echo=FALSE,results = "asis"}
x <- 
  ggplot(
    utmd_output_hello_work_data_yearly,
    aes(
      x = time,
      y = `hire`/1000000,
      group = 1
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  # geom_point(
  #   aes(
  #     shape = industry_group_name,
  #     color = industry_group_name
  #     )
  #   ) +
  theme_classic() + 
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  theme(axis.text.x = element_text(angle=60, vjust=1, hjust=1, size=3.0)) +
  xlab("") +
  ylab("Count (million)") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/hire_year.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)

```

### Beveridge Curve

```{r,echo=FALSE,results = "asis"}
x <- 
  ggplot(
    utmd_output_hello_work_data_yearly,
     aes(
       x = unemployed/1000000,
       y = vacancies/1000000
       )
     ) +
  geom_line(size = 0.3) +
  geom_point(size = 0.3, alpha = 0.3) +
  geom_text(aes(label = time), hjust = 0.5, vjust = -0.5, size = 1.5) +
  theme_classic() +
  labs(
    # title = "Beveridge Curve",
    x = "U (mil)",
    y = "V (mil)"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.caption = element_text(hjust = 0, size = 8)
  ) +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/unemployed_vacancies_berveridge_year.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```

### Job finding rate, worker finding rate

```{r,echo=FALSE,results = "asis"}
temp <-
  utmd_output_hello_work_data_yearly %>% 
  dplyr::mutate(
    `M/U` =
      `hire`/`unemployed`,
    `M/V` =
      `hire`/`vacancies`
  ) %>% 
  dplyr::select(
    `M/U`,
    `M/V`,
    time
  ) %>% 
  tidyr::gather(
    type,
    count,
    - time
  )
x <- 
  ggplot(
    temp,
    aes(
      x = time,
      y = count,
      color = type, 
      group = type
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  theme_classic() + 
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  theme(axis.text.x = element_text(angle=60, vjust=1, hjust=1, size=3.0)) +
  xlab("") +
  ylab("") +
  labs(colour = "") +
  #ylab("Count (million)") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/job_finding_rate_worker_finding_rate_year.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```



### Matching Efficiency A

```{r,echo=FALSE,results = "asis"}
x <- 
  ggplot(
    utmd_output_hello_work_data_yearly,
    aes(
      x = time,
      y = efficiency_implied,
      group = 1
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  # geom_point(
  #   aes(
  #     shape = industry_group_name,
  #     color = industry_group_name
  #     )
  #   ) +
  geom_hline(yintercept = 100, linetype = "longdash") +
  theme_classic() + 
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8))+
  theme(axis.text.x = element_text(angle=60, vjust=1, hjust=1, size=3.0)) +
  xlab("") +
  ylab("A (normalized to 1972)") +
  geom_vline(xintercept = "1972", linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/matching_efficiency_year.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```

### Elasticity dlogM/dlogAU,dlogM/dlogV

```{r,echo=FALSE,results = "asis"}
temp <-
  utmd_output_hello_work_data_yearly %>% 
  dplyr::select(
    `dlog(M)/ dlog(V)`,
    `dlog(M)/ dlog(AU)`,
    time
  ) %>% 
  tidyr::gather(
    type,
    count,
    - time
  )
x <- 
  ggplot(
    temp,
    aes(
      x = time,
      y = count,
      color = type, 
      group = type
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  # geom_point(
  #   aes(
  #     shape = industry_group_name,
  #     color = industry_group_name
  #     )
  #   ) +
  #geom_hline(yintercept = 100, linetype = "longdash") +
  theme_classic() + 
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8))+
  theme(axis.text.x = element_text(angle=60, vjust=1, hjust=1, size=3.0)) +
  xlab("") +
  ylab("log(M) w.r.t. log(AU) and log(V)") +
  labs(colour = "") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/elasticity_year.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```




### The labor market tightness ($V/U$) and Efficiency ($A$)

```{r,echo=FALSE,results = "asis"}
x <-
  ggplot(
    utmd_output_hello_work_data_yearly,
     aes(
       x = log(vacancies/unemployed),
       y = efficiency_implied
       )
     ) +
  #geom_line(size = 0.3) +
  geom_point() +
  #geom_text(aes(label = time), hjust = 0.5, vjust = -0.5, size = 3) +
  theme_classic() +
  labs(
    #title = "Beveridge Curve",
    x = "log(V/U)",
    y = "A"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.caption = element_text(hjust = 0, size = 8)
  ) +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/efficiency_tightness_plot_year.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```

### The job finding rate ($H/U$) , filling rate ($H/V$) and Efficiency ($A$)

```{r,echo=FALSE,results = "asis"}
temp <-
  utmd_output_hello_work_data_yearly %>% 
  dplyr::mutate(
    `log(M/U)` =
      log(`hire`/`unemployed`),
    `log(M/V)` =
      log(`hire`/`vacancies`)
  ) %>% 
  dplyr::select(
    `log(M/U)`,
    `log(M/V)`,
    efficiency_implied
  ) %>% 
  tidyr::gather(
    type,
    count,
    - efficiency_implied
  )

x <-
  ggplot(
    temp,
    aes(
      x = count,
      y = efficiency_implied,
      color = type, 
      group = type
      )
    ) +
  #geom_line(size = 0.3) +
  geom_point() +
  #geom_text(aes(label = time), hjust = 0.5, vjust = -0.5, size = 3) +
  theme_classic() +
  labs(
    #title = "Beveridge Curve",
    x = "log(M/U), log(M/V)",
    y = "A"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.caption = element_text(hjust = 0, size = 8)
  ) +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  labs(colour = "") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/job_finding_rate_efficiency_plot_year.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```




# Appendix: Separate estimation of full-time and part-time {.tabset}

## Residual plot  {.tabset}

### Month-level full-time

```{r,echo=FALSE,results = "asis"}
x <-
  ggplot(
    utmd_output_hello_work_data_full_time_monthly,
     aes(
       x = residual_v_on_u,
       y = residual_a_on_u
       )
     ) +
  #geom_line(size = 0.3) +
  geom_point() +
  #geom_text(aes(label = time), hjust = 0.5, vjust = -0.5, size = 3) +
  theme_classic() +
  labs(
    #title = "Beveridge Curve",
    x = "Residual of V on U",
    y = "Residual of A on U"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.caption = element_text(hjust = 0, size = 8)
  ) +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/residual_plot_month_full_time.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```

### Month-level part-time

```{r,echo=FALSE,results = "asis"}
x <-
  ggplot(
    utmd_output_hello_work_data_part_time_monthly,
     aes(
       x = residual_v_on_u,
       y = residual_a_on_u
       )
     ) +
  #geom_line(size = 0.3) +
  geom_point() +
  #geom_text(aes(label = time), hjust = 0.5, vjust = -0.5, size = 3) +
  theme_classic() +
  labs(
    #title = "Beveridge Curve",
    x = "Residual of V on U",
    y = "Residual of A on U"
  ) +
  theme(
    plot.title = element_text(hjust = 0.5),
    plot.caption = element_text(hjust = 0, size = 8)
  ) +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/residual_plot_month_part_time.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```
