==> main/05_construct_figuretable_prefecture_job_category.Rmd <==
---
title: "05_construct_figuretable_prefecture_job_category"
author: "Suguru Otani"
date: "2024-01-26"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# load {.tabset}

```{r, echo=FALSE,results = "asis"}
rm(list = ls())
library(magrittr)
library(ggplot2)
library(modelsummary) # for N in datasummary

# utmd_output_hello_work_data_part_time_monthly_prefecture <-
#   readRDS(file = here::here("output/utmd_output_hello_work_data_part_time_monthly_prefecture.rds")) %>% 
#   dplyr::rename(
#     `unemployed`= candidate_count,
#     `vacancies` = position_count,
#     `hire` = hire_count,
#     `dlog(M)/ dlog(V)` = hire_elasticity_vacancy,
#     `dlog(M)/ dlog(AU)` = hire_elasticity_efficiency_unemployed
#   )

# utmd_output_hello_work_data_full_time_monthly_prefecture <-
#   readRDS(file = here::here("output/utmd_output_hello_work_data_full_time_monthly_prefecture.rds")) %>% 
#   dplyr::rename(
#     `unemployed`= candidate_count,
#     `vacancies` = position_count,
#     `hire` = hire_count,
#     `dlog(M)/ dlog(V)` = hire_elasticity_vacancy,
#     `dlog(M)/ dlog(AU)` = hire_elasticity_efficiency_unemployed
#   )
job_category_library <-
  readRDS(file = here::here("cleaned/job_category_library.rds"))
prefecture_library <-
  readRDS(file = here::here("cleaned/prefecture_library.rds"))
utmd_output_hello_work_data_part_and_full_time_monthly_prefecture <-
  readRDS(file = here::here("output/utmd_output_hello_work_data_part_and_full_time_monthly_prefecture.rds")) %>% 
  dplyr::rename(
    `unemployed`= candidate_count,
    `vacancies` = position_count,
    `hire` = hire_count,
    `dlog(M)/ dlog(V)` = hire_elasticity_vacancy,
    `dlog(M)/ dlog(AU)` = hire_elasticity_efficiency_unemployed
  ) %>% 
  dplyr::left_join(
    prefecture_library,
    by = c("prefecture" = "prefecture")
  )
utmd_output_hello_work_data_part_and_full_time_monthly_job_category <-
  readRDS(file = here::here("output/utmd_output_hello_work_data_part_and_full_time_monthly_job_category.rds")) %>% 
  dplyr::rename(
    `unemployed`= candidate_count,
    `vacancies` = position_count,
    `hire` = hire_count,
    `dlog(M)/ dlog(V)` = hire_elasticity_vacancy,
    `dlog(M)/ dlog(AU)` = hire_elasticity_efficiency_unemployed
  ) %>% 
  dplyr::left_join(
    job_category_library,
    by = c("job_kinds" = "job_kinds")
  )

utmd_output_hello_work_data_part_and_full_time_monthly_optimal_hire_prefecture <-
  readr::read_csv(file = here::here("output/utmd_output_hello_work_data_part_and_full_time_monthly_optimal_hire_prefecture.csv"))[-1,]
colnames(utmd_output_hello_work_data_part_and_full_time_monthly_optimal_hire_prefecture) <-
  c(
    "time",
    "prefecture",
    "optimal_unemployed",
    "optimal_hire"
  )
utmd_output_hello_work_data_part_and_full_time_monthly_optimal_hire_prefecture <-
  utmd_output_hello_work_data_part_and_full_time_monthly_optimal_hire_prefecture %>% 
  dplyr::mutate(
    time = as.Date(time),
    optimal_hire = as.numeric(optimal_hire)
  )
utmd_output_hello_work_data_part_and_full_time_monthly_optimal_hire_job_category <-
  readr::read_csv(file = here::here("output/utmd_output_hello_work_data_part_and_full_time_monthly_optimal_hire_job_category.csv"))[-1,]
colnames(utmd_output_hello_work_data_part_and_full_time_monthly_optimal_hire_job_category) <-
  c(
    "time",
    "job_kinds",
    "optimal_unemployed",
    "optimal_hire"
  )
utmd_output_hello_work_data_part_and_full_time_monthly_optimal_hire_job_category <-
  utmd_output_hello_work_data_part_and_full_time_monthly_optimal_hire_job_category %>% 
  dplyr::mutate(
    time = as.Date(time),
    optimal_hire = as.numeric(optimal_hire)
  )
utmd_output_hello_work_data_part_and_full_time_monthly_optimal_hire_job_category <-
  utmd_output_hello_work_data_part_and_full_time_monthly_optimal_hire_job_category %>% 
  dplyr::mutate(
    time = as.Date(time),
    optimal_hire = as.numeric(optimal_hire)
  )

hokkaido = c("Hokkaido")
tohoku = c("Aomori", "Iwate", "Miyagi", "Akita", "Yamagata", "Fukushima")
kanto = c("Ibaraki", "Tochigi", "Gunma", "Saitama", "Chiba", "Tokyo", "Kanagawa")
chubu = c("Niigata", "Toyama", "Ishikawa", "Fukui", "Yamanashi", "Nagano", "Gifu", "Shizuoka", "Aichi")
kansai = c("Mie", "Shiga", "Kyoto", "Osaka", "Hyogo", "Nara", "Wakayama")
chugoku = c("Tottori", "Shimane", "Okayama", "Hiroshima", "Yamaguchi")
shikoku = c("Tokushima", "Kagawa", "Ehime", "Kochi")
kyushu = c("Fukuoka", "Saga", "Nagasaki", "Kumamoto", "Oita", "Miyazaki", "Kagoshima", "Okinawa")

managerial = c("Managerial")
professional_and_technical = c("Prof. & Tech.", "Dev. Eng.", "Mfg. Eng.", 
                                 "Arch. & Civil Eng.", "IT & Comm. Eng.", 
                                 "Other Eng.", "Doctors, etc.", "Nurses", 
                                 "Med. Tech.", "Other Health", "Soc. Welfare", 
                                 "Artists, etc.", "Other Prof.")
clerical = c("Clerical", "Gen. Clerical", "Acct. Clerical", "Prod. Clerical", 
               "Sales Clerical", "Field Clerical", "Trans. Clerical", "Office Ops")
sales = c("Sales", "Merch. Sales", "Sales Rel.", "Sales Reps")
service = c("Service", "Home Support", "Care Service", "Health Service", 
              "Sanitation", "Food Prep.", "Reception", "Bldg. Mgmt.", "Other Service")
agriculture_forestry_and_fishing = c("Agri. & Fish.")
production = c("Production", "Prod. Ctrl (Met.)", "Prod. Ctrl (Non-Met.)", 
                 "Prod. Ctrl (Mach.)", "Metalwork", "Prod. (Non-Met.)", 
                 "Mach. Assembly", "Mach. Maint.", "Prod. Insp. (Met.)", 
                 "Prod. Insp. (Non-Met.)", "Mach. Insp.", "Prod. Rel.")
transportation_and_machine_operation = c("Trans. Ops", "Railway Ops", "Auto Ops", 
                                           "Ship & Air Ops", "Other Trans.", 
                                           "Const. Mach. Ops")
construction_and_mining = c("Construction", "Frame Const.", "Const. Work", 
                              "Elec. Const.", "Civil Eng.", "Mining")
transportation_cleaning_and_packaging = c("Trans. & Clean.", "Transport", 
                                            "Cleaning", "Packaging", 
                                            "Other Trans. & Clean.")
security = c("Security")
# others = c("Care")
# others = c("介護関係職種（注２）")

generate_efficiency_plot <-
  function(
    target_data,
    target_area,
    cross_sectional_target
    ){
    if(cross_sectional_target == "prefecture"){
      target_data <-
        target_data %>% 
        dplyr::rename(
          target_var = prefecture_english
        )
      target_y_label <-
        "A (normalized to 2013 Jan Tokyo)"
      y_upper_bound <-
        600
    } else if(cross_sectional_target == "job_category"){
      target_data <-
        target_data %>% 
        dplyr::rename(
          target_var = job_kinds_english
        )
      target_y_label <-
        "A (normalized to 2013 Jan Clerical)"
      y_upper_bound <-
        1000
    }
    x <- 
      ggplot(
        target_data %>% 
          dplyr::filter(
            target_var %in% target_area
          ),
        aes(
          x = time,
          y = efficiency_implied,
          color = target_var, 
          group = target_var
          )
        ) +
      geom_line(
        alpha = 0.8, 
        size = 0.3
        ) + 
      # geom_point(
      #   aes(
      #     shape = industry_group_name,
      #     color = industry_group_name
      #     )
      #   ) +
      geom_hline(yintercept = 100, linetype = "longdash") +
      theme_classic() + 
      #scale_color_viridis_d() +
      theme(legend.position = 'top',
            legend.text = element_text(size=6))+
      #theme(axis.text.x = element_text(angle=60, vjust=1, hjust=1, size=3.0)) +
      xlab("") +
      ylim(0, y_upper_bound) +
      ylab(target_y_label) +
      labs(colour = "") +
      #geom_vline(xintercept = 2002, linetype = "longdash") +
      # geom_vline(xintercept = 1984, linetype = "longdash") +
      scale_color_viridis_d()
    x
    figure_name <-
      paste(
        "../figuretable/matching_efficiency_month_aggregate_",
        deparse(substitute(target_area)),
        ".png",
        sep = ""
        )
    ggsave(filename = figure_name,
           plot = x,
           device = "png",
           width = 4,
           height = 3)
    return(x)
  }

generate_elasticity_plot <-
  function(
    target_data,
    target_area,
    cross_sectional_target
    ){
    if(cross_sectional_target == "prefecture"){
      target_data <-
        target_data %>% 
        dplyr::rename(
          target_var = prefecture_english
        )
      y_upper_bound_elasticity_unemployed <-
        1.4
      y_upper_bound_elasticity_vacancy <-
        0.8
    } else if(cross_sectional_target == "job_category"){
      target_data <-
        target_data %>% 
        dplyr::rename(
          target_var = job_kinds_english
        )
      y_upper_bound_elasticity_unemployed <-
        1.8
      y_upper_bound_elasticity_vacancy <-
        6.0
    }
    # elastiticy w.r.t. vacancy
    x <- 
      ggplot(
        target_data %>% 
          dplyr::filter(
            target_var %in% target_area
          ),
        aes(
          x = time,
          y = `dlog(M)/ dlog(V)`,
          color = target_var, 
          group = target_var
          )
        ) +
      geom_line(
        alpha = 0.8, 
        size = 0.3
        ) + 
      # geom_point(
      #   aes(
      #     shape = industry_group_name,
      #     color = industry_group_name
      #     )
      #   ) +
      #geom_hline(yintercept = 100, linetype = "longdash") +
      theme_classic() + 
      #scale_color_viridis_d() +
      theme(legend.position = 'top',
            legend.text = element_text(size=6))+
      xlab("") +
      ylab("log(M) w.r.t. log(V)") +
      ylim(0, y_upper_bound_elasticity_vacancy) +
      labs(colour = "") +
      # geom_vline(xintercept = 1980, linetype = "longdash") +
      # geom_vline(xintercept = 1984, linetype = "longdash") +
      scale_color_viridis_d()
    figure_name <-
      paste(
        "../figuretable/elasticity_vacancy_month_aggregate_",
        deparse(substitute(target_area)),
        ".png",
        sep = ""
        )
    ggsave(filename = figure_name,
           plot = x,
           device = "png",
           width = 4,
           height = 3)
    # elastiticy w.r.t. AU
    y <- 
      ggplot(
        target_data %>% 
          dplyr::filter(
            target_var %in% target_area
          ),
        aes(
          x = time,
          y = `dlog(M)/ dlog(AU)`,
          color = target_var, 
          group = target_var
          )
        ) +
      geom_line(
        alpha = 0.8, 
        size = 0.3
        ) + 
      # geom_point(
      #   aes(
      #     shape = industry_group_name,
      #     color = industry_group_name
      #     )
      #   ) +
      #geom_hline(yintercept = 100, linetype = "longdash") +
      theme_classic() + 
      #scale_color_viridis_d() +
      theme(legend.position = 'top',
            legend.text = element_text(size=6))+
      xlab("") +
      ylab("log(M) w.r.t. log(AU)") +
      labs(colour = "") +
      ylim(0, y_upper_bound_elasticity_unemployed) +
      # geom_vline(xintercept = 1980, linetype = "longdash") +
      # geom_vline(xintercept = 1984, linetype = "longdash") +
      scale_color_viridis_d()
    figure_name <-
      paste(
        "../figuretable/elasticity_unemployed_month_aggregate_",
        deparse(substitute(target_area)),
        ".png",
        sep = ""
        )
    ggsave(filename = figure_name,
           plot = y,
           device = "png",
           width = 4,
           height = 3)
    res <-
      list(
        elasticity_vacancy = x,
        elasticity_unemployed = y
        )
    return(res)
  }
```



## Figure 6: Month-prefecture-level aggregate results 2012-2024 {.tabset}

the match elasticity with respect to unemployment is in the range 0.5–0.7.

### Matching Efficiency A

```{r,echo=FALSE,results = "asis"}
cross_sectional_target <-
  "prefecture"
target_data <-
  utmd_output_hello_work_data_part_and_full_time_monthly_prefecture
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = hokkaido,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = tohoku,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = kanto,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = chubu,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = kansai,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = chugoku,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = shikoku,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = kyushu,
    cross_sectional_target
    )
x

```

### Elasticity dlogM/dlogU,dlogM/dlogV

```{r,echo=FALSE,results = "asis"}
cross_sectional_target <-
  "prefecture"
target_data <-
  utmd_output_hello_work_data_part_and_full_time_monthly_prefecture
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = hokkaido,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = tohoku,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = kanto,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = chubu,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = kansai,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = chugoku,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = shikoku,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = kyushu,
    cross_sectional_target
    )
x

```


## Figure 7: Month-job-category-level aggregate trends in 2012-2024 {.tabset}

### Matching Efficiency A

```{r,echo=FALSE,results = "asis"}
cross_sectional_target <-
  "job_category"
target_data <-
  utmd_output_hello_work_data_part_and_full_time_monthly_job_category
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = managerial,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = professional_and_technical,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = clerical,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = sales,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = service,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = agriculture_forestry_and_fishing,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = production,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = transportation_and_machine_operation,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = construction_and_mining,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = transportation_cleaning_and_packaging,
    cross_sectional_target
    )
x
x <- 
  generate_efficiency_plot(
    target_data,
    target_area = security,
    cross_sectional_target
    )
x


```



### Elasticity dlogM/dlogU,dlogM/dlogV

```{r,echo=FALSE,results = "asis"}
cross_sectional_target <-
  "job_category"
target_data <-
  utmd_output_hello_work_data_part_and_full_time_monthly_job_category
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = managerial,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = professional_and_technical,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = clerical,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = sales,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = service,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = agriculture_forestry_and_fishing,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = production,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = transportation_and_machine_operation,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = construction_and_mining,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = transportation_cleaning_and_packaging,
    cross_sectional_target
    )
x
x <- 
  generate_elasticity_plot(
    target_data,
    target_area = security,
    cross_sectional_target
    )
x
```

## Figure 8: Mismatch 1-H/H* {.tabset}

### Across prefectures

```{r,echo=FALSE,results = "asis"}
baseline_mismatch_temp <-
  0.1
taget_data <-
  utmd_output_hello_work_data_part_and_full_time_monthly_prefecture  %>% 
  dplyr::left_join(
    utmd_output_hello_work_data_part_and_full_time_monthly_optimal_hire_prefecture,
    by = 
      c("prefecture" = "prefecture",
        "time" = "time")
  ) %>% 
  dplyr::group_by(
    time
    ) %>% 
  dplyr::summarise(
    total_optimal_hire = sum(optimal_hire),
    total_actual_hire = sum(hire)
  ) %>% 
  dplyr::ungroup() %>% 
  dplyr::mutate(
    mismatch =
      (1 - (total_actual_hire/total_optimal_hire)) 
    + baseline_mismatch_temp
  )


x <- 
  ggplot(
    taget_data,
    aes(
      x = time,
      y = mismatch
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  # geom_point(
  #   aes(
  #     shape = industry_group_name,
  #     color = industry_group_name
  #     )
  #   ) +
  theme_classic() + 
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  xlab("") +
  ylab("Mismatch across prefectures") +
  labs(colour = "") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/mismatch_part_and_full_time_monthly_prefecture.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```

### Across job categories

```{r,echo=FALSE,results = "asis"}
taget_data <-
  utmd_output_hello_work_data_part_and_full_time_monthly_job_category %>% 
  dplyr::left_join(
    utmd_output_hello_work_data_part_and_full_time_monthly_optimal_hire_job_category,
    by = 
      c("job_kinds" = "job_kinds",
        "time" = "time")
  ) %>% 
  dplyr::group_by(
    time
    ) %>% 
  dplyr::summarise(
    total_optimal_hire = sum(optimal_hire),
    total_actual_hire = sum(hire)
  ) %>% 
  dplyr::ungroup() %>% 
  dplyr::mutate(
    mismatch = 1 - (total_actual_hire/total_optimal_hire)
  ) %>% 
  dplyr::filter(total_optimal_hire >= total_actual_hire)
x <- 
  ggplot(
    taget_data,
    aes(
      x = time,
      y = mismatch
      )
    ) +
  geom_line(
    alpha = 0.8, 
    size = 0.3
    ) + 
  # geom_point(
  #   aes(
  #     shape = industry_group_name,
  #     color = industry_group_name
  #     )
  #   ) +
  theme_classic() + 
  #scale_color_viridis_d() +
  theme(legend.position = 'top',
        legend.text = element_text(size=8)) +
  xlab("") +
  ylab("Mismatch across job categories") +
  labs(colour = "") +
  # geom_vline(xintercept = 1980, linetype = "longdash") +
  # geom_vline(xintercept = 1984, linetype = "longdash") +
  scale_color_viridis_d()
x
figure_name <-
  "../figuretable/mismatch_part_and_full_time_monthly_job_category.png"
ggsave(filename = figure_name,
       plot = x,
       device = "png",
       width = 4,
       height = 3)
```

# Appendix {.tabset}

## Table A1: Month-job-category-level aggregate trends in 2012-2024 {.tabset}

### Job category list

```{r,echo=FALSE,results = "asis"}
job_category_library %>%
  dplyr::select(
    job_kinds_english,
    job_kinds
  ) %>% 
  dplyr::rename(
    `English` = job_kinds_english,
    `Japanese` = job_kinds
  ) %>% 
  kableExtra::kbl(
    format = "latex",
    booktabs = TRUE,
    escape = FALSE,
    linesep = "",
    align = "ll"
  ) %>%
  kableExtra::save_kable(
    file = here::here("figuretable/job_category_list.tex")
  )
job_category_library %>% 
  dplyr::select(
    job_kinds_english,
    job_kinds
  ) %>% 
  dplyr::rename(
    `English` = job_kinds_english,
    `Japanese` = job_kinds
  ) %>% 
  kableExtra::kable() %>% 
  kableExtra::kable_styling()
```
