==> main/06_estimate_simulation_data.Rmd <==
---
title: "03_construct_figuretable"
author: "Suguru Otani"
date: "2024-01-26"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# load {.tabset}

```{r, echo=FALSE,results = "asis"}
rm(list = ls())
library(magrittr)
library(ggplot2)
# set constant ----
list_num_time <-
  c(
    #10,
    #20,
    #30,
    #40,
    50,
    100,
    200
  )
specification_list <-
  c(
    "cobb_douglas",
    "perfect_substitute"#,
    #"fixed_proportion"
  )
CRS_gamma_parameter <-
  0.3#0.001
arima_list <-
  list(
    "AR1_I0" = c(1,0,0),
    #"AR1_I1" = c(1,1,0),
    "AR0_I1" = c(0,1,0)
  )
for(nn in 1:length(list_num_time)){
  for(mm in 1:length(specification_list)){
    for(rr in 1:length(arima_list)){
      target_arima <-
        arima_list[[rr]]
      target_arima_name <-
        names(arima_list)[rr]
      CRS_gamma_parameter <-
      0.3
    temp_nn <-
      list_num_time[nn]
    matching_function_specification <-
      specification_list[mm]
    temp_nn <-
      list_num_time[nn]
    filename <-
      paste(
        #"output/monte_carlo_data_",
        "num_time_",
        temp_nn,
        "_",
        matching_function_specification,
        "_",
        CRS_gamma_parameter,
        "_",
        target_arima_name,
        sep = ""
      )
    cat(filename,"\n")
    temp_data <-
      readRDS(
        file =
          here::here(
            paste(
              "output/",
              "implied_efficiency_",
              filename,
              ".rds",
              sep = ""
            )
            )
        )
    assign(
      filename,
      temp_data)
      }
    
  }
}
generate_bias_rmse <-
  function(target_data){
    res <-
      target_data %>% 
      dplyr::mutate(
        error = normalized_efficiency - efficiency_implied
        ) %>% 
      dplyr::group_by(
        simulation_id
      ) %>% 
      dplyr::summarise(
        bias =
          mean(error),
        rmse =
          sqrt(mean((error)^2, na.rm = TRUE))
      ) %>% 
      dplyr::ungroup()
    return(res)
  }

options(modelsummary_format_numeric_latex = "plain") # for modelsummary output
```



## Stationary cobb_douglas {.tabset}

the match elasticity with respect to unemployment is in the range 0.5–0.7.

### num_time_50_cobb_douglas_0.3_AR1_I0

```{r,echo=FALSE,results = "asis"}
target_res <-
  generate_bias_rmse(num_time_50_cobb_douglas_0.3_AR1_I0)
modelsummary::datasummary_skim(
  fmt = 3,
  target_res %>% 
    dplyr::filter(bias <= quantile(bias, 0.9))
  )
```

### num_time_100_cobb_douglas_0.3_AR1_I0

```{r,echo=FALSE,results = "asis"}
target_res <-
  generate_bias_rmse(num_time_100_cobb_douglas_0.3_AR1_I0)
modelsummary::datasummary_skim(
  fmt = 3,
  target_res %>% 
    dplyr::filter(bias <= quantile(bias, 0.9))
  )
```


### num_time_200_cobb_douglas_0.3_AR1_I0

```{r,echo=FALSE,results = "asis"}
target_res <-
  generate_bias_rmse(num_time_200_cobb_douglas_0.3_AR1_I0)
modelsummary::datasummary_skim(
  fmt = 3,
  target_res %>% 
    dplyr::filter(bias <= quantile(bias, 0.9))
  )
```


## Stationary perfect_substitute {.tabset}

the match elasticity with respect to unemployment is in the range 0.5–0.7.

### num_time_50_perfect_substitute_0.3_AR1_I0

```{r,echo=FALSE,results = "asis"}
target_res <-
  generate_bias_rmse(num_time_50_perfect_substitute_0.3_AR1_I0)
modelsummary::datasummary_skim(
  fmt = 3,
  target_res %>% 
    dplyr::filter(bias <= quantile(bias, 0.9))
  )
```

### num_time_100_perfect_substitute_0.3_AR1_I0

```{r,echo=FALSE,results = "asis"}
target_res <-
  generate_bias_rmse(num_time_100_perfect_substitute_0.3_AR1_I0)
modelsummary::datasummary_skim(
  fmt = 3,
  target_res %>% 
    dplyr::filter(bias <= quantile(bias, 0.9))
  )
```


### num_time_200_perfect_substitute_0.3_AR1_I0

```{r,echo=FALSE,results = "asis"}
target_res <-
  generate_bias_rmse(num_time_200_perfect_substitute_0.3_AR1_I0)
modelsummary::datasummary_skim(
  fmt = 3,
  target_res %>% 
    dplyr::filter(bias <= quantile(bias, 0.9))
  )
```

## Nonstationary cobb_douglas {.tabset}

the match elasticity with respect to unemployment is in the range 0.5–0.7.

### num_time_50_cobb_douglas_0.3_AR0_I1

```{r,echo=FALSE,results = "asis"}
target_res <-
  generate_bias_rmse(num_time_50_cobb_douglas_0.3_AR0_I1)
modelsummary::datasummary_skim(
  fmt = 3,
  target_res %>% 
    dplyr::filter(bias <= quantile(bias, 0.9))
  )
```

### num_time_100_cobb_douglas_0.3

```{r,echo=FALSE,results = "asis"}
target_res <-
  generate_bias_rmse(num_time_100_cobb_douglas_0.3_AR0_I1)
modelsummary::datasummary_skim(
  fmt = 3,
  target_res %>% 
    dplyr::filter(bias <= quantile(bias, 0.9))
  )
```


### num_time_200_cobb_douglas_0.3

```{r,echo=FALSE,results = "asis"}
target_res <-
  generate_bias_rmse(num_time_200_cobb_douglas_0.3_AR0_I1)
modelsummary::datasummary_skim(
  fmt = 3,
  target_res %>% 
    dplyr::filter(bias <= quantile(bias, 0.9))
  )
```


## Nonstationary perfect_substitute {.tabset}

the match elasticity with respect to unemployment is in the range 0.5–0.7.

### num_time_50_perfect_substitute_0.3

```{r,echo=FALSE,results = "asis"}
target_res <-
  generate_bias_rmse(num_time_50_perfect_substitute_0.3_AR0_I1)
modelsummary::datasummary_skim(
  fmt = 3,
  target_res %>% 
    dplyr::filter(bias <= quantile(bias, 0.9))
  )
```

### num_time_100_perfect_substitute_0.3

```{r,echo=FALSE,results = "asis"}
target_res <-
  generate_bias_rmse(num_time_100_perfect_substitute_0.3_AR0_I1)
modelsummary::datasummary_skim(
  fmt = 3,
  target_res %>% 
    dplyr::filter(bias <= quantile(bias, 0.9))
  )
```


### num_time_200_perfect_substitute_0.3

```{r,echo=FALSE,results = "asis"}
target_res <-
  generate_bias_rmse(num_time_200_perfect_substitute_0.3_AR0_I1)
modelsummary::datasummary_skim(
  fmt = 3,
  target_res %>% 
    dplyr::filter(bias <= quantile(bias, 0.9))
  )
```
